v0.5.0 [General] - Added email verification for users upon first login - Added responsiveness to Add Beneficiaries Modal - Added SMS verification for every login (temporarily disabled) - User now cannot add a "special case" beneficiary when the original beneficiary is still under pending implementation - Added a logic where it deletes old images when updating new images to save server memory - Reworked the Import functionality with error validation, edit & delete functions, and better indicators - Reworked the Sample Format to make it more eloquent to work with - Fixed a bug when adding a special case without images - Fixed the color indicators for similarity table on Add & Edit Beneficiary Modals (focal & coordinator) - Fixed algorithm causing errors when checking names with symbols - Fixed special case description not triggering validation when it is empty and not null - Fixed Delete prompt `Cancel` not closing the modal - Fixed export on summary being able to export without any beneficiaries - Fixed annex export error when changing filter dates - Fixed some image errors on Add & Edit modals
[Focal]
- Added function to change approval status (force approve & pend batch)
- Added a static online indicator in User Management
- Improved the UI for Export Summary Modal & Added extra info for selected project
- Fixed Assign Batches not refreshing Add Coordinator list after adding a batch row
- Fixed batch number resetting to null when adding a batch row

[Coordinator]
- Added "No Other Coordinator" indicator for solo coordinators on a batch

[Misc]
- Capitalized the spouse names
- Centered the login page floater box

+++ app/Livewire/Barangay/ListingPage.php
@@ -345,6 +345,8 @@ class ListingPage extends Component
     #[On('edit-beneficiary')]
     public function editBeneficiary()
     {
+        $this->beneficiaryId = null;
+        $this->selectedBeneficiaryRow = -1;
         $this->showAlert = true;
         $this->alertMessage = 'Successfully modified a beneficiary!';
         $this->dispatch('show-alert');
@@ -362,7 +364,7 @@ class ListingPage extends Component
                 ->get();
             $old = $beneficiary;
             foreach ($credentials as $credential) {
-                if (isset($credential->image_file_path)) {
+                if (isset($credential->image_file_path) && Storage::exists($credential->image_file_path)) {
                     Storage::delete($credential->image_file_path);
                 }
                 $credential->delete();
